Commented out auth guards temporarily (for development communication tests). Validations added and comments

+++ libs/auth-api-lib/src/lib/services/resources.service.ts
@@ -1,4 +1,4 @@
-import { Inject, Injectable } from '@nestjs/common'
+import { BadRequestException, Inject, Injectable } from '@nestjs/common'
 import { InjectModel } from '@nestjs/sequelize'
 import { Logger, LOGGER_PROVIDER } from '@island.is/logging'
 import { Op, WhereOptions } from 'sequelize'
@@ -28,26 +28,41 @@ export class ResourcesService {
     private logger: Logger,
   ) {}
 
+  /** Gets Identity resource by id */
   async getIdentityResourceById(id: string): Promise<IdentityResource> {
     this.logger.debug('Getting data about identity resource with id: ', id)
+
+    if (!id) {
+      throw new BadRequestException('Id must be provided')
+    }
 
     return this.identityResourceModel.findOne({
       where: { id: id },
     })
   }
 
+  /** Gets API scope by id */
   async getApiScopeById(id: string): Promise<ApiScope> {
     this.logger.debug('Getting data about api scope with id: ', id)
+
+    if (!id) {
+      throw new BadRequestException('Id must be provided')
+    }
 
     return this.apiScopeModel.findOne({
       where: { id: id },
     })
   }
 
+  /** Get identity resources by scope names */
   async findIdentityResourcesByScopeName(
     scopeNames: string[],
   ): Promise<IdentityResource[]> {
     this.logger.debug(`Finding identity resources for scope names`, scopeNames)
+
+    if (!scopeNames) {
+      throw new BadRequestException('ScopeNames must be provided')
+    }
 
     const whereOptions: WhereOptions = {
       name: {
@@ -61,8 +76,13 @@ export class ResourcesService {
     })
   }
 
+  /** Gets Api scopes by scope names  */
   async findApiScopesByNameAsync(scopeNames: string[]): Promise<ApiScope[]> {
     this.logger.debug(`Finding api scopes for scope names`, scopeNames)
+
+    if (!scopeNames) {
+      throw new BadRequestException('ScopeNames must be provided')
+    }
 
     const whereOptions: WhereOptions = {
       name: {
@@ -76,6 +96,7 @@ export class ResourcesService {
     })
   }
 
+  /** Gets api resources by api resource names  */
   async findApiResourcesByNameAsync(
     apiResourceNames: string[],
   ): Promise<ApiResource[]> {
@@ -83,6 +104,10 @@ export class ResourcesService {
       `Finding api resources for resource names`,
       apiResourceNames,
     )
+
+    if (!apiResourceNames) {
+      throw new BadRequestException('ApiResourceNames must be provided')
+    }
 
     const whereOptions: WhereOptions = {
       name: {
@@ -96,6 +121,7 @@ export class ResourcesService {
     })
   }
 
+  /** Get Api resources by scope names */
   async findApiResourcesByScopeNameAsync(
     apiResourceScopeNames: string[],
   ): Promise<ApiResource[]> {
@@ -103,6 +129,10 @@ export class ResourcesService {
       `Finding api resources for resource scope names`,
       apiResourceScopeNames,
     )
+
+    if (!apiResourceScopeNames) {
+      throw new BadRequestException('ApiResourceNames must be provided')
+    }
 
     const scopesWhereOptions: WhereOptions = {
       scopeName: {
@@ -125,6 +155,7 @@ export class ResourcesService {
     })
   }
 
+  /** Creates a new identity resource */
   async createIdentityResource(
     identityResource: IdentityResourcesDTO,
   ): Promise<IdentityResource> {
@@ -133,11 +164,16 @@ export class ResourcesService {
     return await this.identityResourceModel.create({ ...identityResource })
   }
 
+  /** Updates an existing Identity resource */
   async updateIdentityResource(
     identityResource: IdentityResourcesDTO,
     id: string,
   ): Promise<IdentityResource> {
     this.logger.debug('Updating identity resource with id: ', id)
+
+    if (!id) {
+      throw new BadRequestException('Id must be provided')
+    }
 
     await this.identityResourceModel.update(
       { ...identityResource },
@@ -147,28 +183,44 @@ export class ResourcesService {
     return await this.getIdentityResourceById(id)
   }
 
+  /** Deletes an identity resource by id */
   async deleteIdentityResource(id: string): Promise<number> {
     this.logger.debug('Removing identity resource with id: ', id)
+
+    if (!id) {
+      throw new BadRequestException('Id must be provided')
+    }
 
     return await this.identityResourceModel.destroy({ where: { id: id } })
   }
 
+  /** Creates a new Api Scope */
   async createApiScope(apiScope: ApiScopesDTO): Promise<ApiScope> {
     this.logger.debug('Creating a new api scope')
 
     return await this.apiScopeModel.create({ ...apiScope })
   }
 
+  /** Updates an existing API scope */
   async updateApiScope(apiScope: ApiScopesDTO, id: string): Promise<ApiScope> {
     this.logger.debug('Updating api scope with id: ', id)
+
+    if (!id) {
+      throw new BadRequestException('Id must be provided')
+    }
 
     await this.apiScopeModel.update({ ...apiScope }, { where: { id: id } })
 
     return this.getApiScopeById(id)
   }
 
+  /** Deletes an API scope */
   async deleteApiScope(id: string): Promise<number> {
     this.logger.debug('Deleting api scope with id: ', id)
+
+    if (!id) {
+      throw new BadRequestException('Id must be provided')
+    }
 
     return await this.apiScopeModel.destroy({ where: { id: id } })
   }
