+++ libs/auth-api-lib/src/lib/services/user-identities.service.ts
@@ -1,4 +1,4 @@
-import { Inject, Injectable } from '@nestjs/common'
+import { BadRequestException, Inject, Injectable } from '@nestjs/common'
 import { InjectModel } from '@nestjs/sequelize'
 import { Logger, LOGGER_PROVIDER } from '@island.is/logging'
 import { Claim } from '../entities/models/claim.model'
@@ -16,6 +16,7 @@ export class UserIdentitiesService {
     private logger: Logger,
   ) {}
 
+  /** Creates a new User Identity */
   async create(userIdentity: UserIdentityDto): Promise<UserIdentity> {
     this.logger.debug(
       `Creating user identity with subjectId - ${userIdentity.subjectId}`,
@@ -33,14 +34,21 @@ export class UserIdentitiesService {
     }
   }
 
+  /** Gets a user identity by subject id */
   async findBySubjectId(subjectId: string): Promise<UserIdentity> {
     this.logger.debug(`Finding user identity for subjectId - "${subjectId}"`)
+
+    if (!subjectId) {
+      throw new BadRequestException('SubjectId must be provided')
+    }
+
     return await this.userIdentityModel.findOne({
       where: { subjectId },
       include: [Claim],
     })
   }
 
+  /** Gets a user identiy by a provider and subjectid */
   async findByProviderSubjectId(
     provider: string,
     subjectId: string,
@@ -48,12 +56,18 @@ export class UserIdentitiesService {
     this.logger.debug(
       `Finding user identity for provider "${provider}" and subjectId - "${subjectId}"`,
     )
+
+    if (!provider) {
+      throw new BadRequestException('Provider must be provided')
+    }
+
     return this.userIdentityModel.findOne({
       where: { providerName: provider, providerSubjectId: subjectId },
       include: [Claim],
     })
   }
 
+  /** Updates an existing user identity */
   async update(userIdentity: UserIdentityDto): Promise<UserIdentity> {
     this.logger.debug(
       'Updating the user identity with id: ',
@@ -70,8 +84,13 @@ export class UserIdentitiesService {
     return await this.findBySubjectId(userIdentity.subjectId)
   }
 
+  /** Deletes an user identity by subjectid */
   async delete(subjectId: string): Promise<number> {
     this.logger.debug('Deleting user identity with id: ', subjectId)
+
+    if (!subjectId) {
+      throw new BadRequestException('SubjectId must be provided')
+    }
 
     return await this.userIdentityModel.destroy({
       where: { subjectId: subjectId },
